+++ CHANGELOG.md
@@ -21,6 +21,7 @@ For notes on migrating to 3.x see [the 3.x migration guide](doc/3.x/migration-gu
 
 ### :bug: Bug Fixes
 
+* fix(sdk-trace): return no-op tracer when TracerProvider is shutdown [#3211](https://github.com/open-telemetry/opentelemetry-js/issues/3211) @shubhamrai9122-creator
 * fix(sdk-metrics): guard against a missing or empty name to MeterProvider#getMeter [#7105](https://github.com/open-telemetry/opentelemetry-js/pull/7105) @trentm
 
 ### :books: Documentation